requirements.txt
# Production dependencies for Vercel deployment
flask>=3.0.0
flask-cors>=4.0.0
tableauserverclient>=0.31
python-dotenv>=1.0.0
requests>=2.31.0
